make popularityScore Integer instead of Double

+++ src/main/java/com/grabduck/githubsearch/api/RepositoryDto.java
@@ -13,5 +13,5 @@ public record RepositoryDto(
     String language,
     Integer stargazersCount,
     Integer forksCount,
-    Double popularityScore
+    Integer popularityScore
 ) {}
